Seperate scrape and tweet-out func into own files, export/import and use from routes
